pyproject.toml
[tool.poetry]
name = "repl_python3_Kur0bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyNaCl = "^1.4.0"
gtts = "^2.2.2"
pet-pet-gif = "^1.0.1"
Flask = "^2.0.1"
python-dateutil = "^2.8.2"
yt-dlp = "^2021.10.10"
mcrcon = "^0.7.0"
saucenao-api = {git = "https://github.com/jericjan/saucenao_api.git"}
aiohttp="<3.8.0"
#py-cord = {git = "https://github.com/Pycord-Development/pycord.git"}
black = "^22.6.0"
pytz = "^2021.3"
disnake = "^2.3.0"
tqdm = "^4.62.3"
arsenic = "^21.8"
Pillow = "^9.0.1"
pilmoji = "^2.0.1"
aiolimiter = "^1.0.0"
lorem = "^0.1.1"
translators = "^5.2.2"
opencv-python = "^4.6.0"
ffmpegcv = "^0.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
